pyproject.toml
[project]
name = "hillbot"
version = "0.1.0"
description = "Personal Discord bot for image generation and chat"
authors = [
    { name = "Cell Webb" }
]
requires-python = ">=3.10"

dependencies = [
    "discord.py>=2.3.0",
    "numpy>=1.24.0",
    "openai>=1.3.0",
    "replicate>=0.10.0",
    "pyyaml>=6.0.1",
    "pillow>=10.0.0",
    "python-dotenv>=1.0.1",
]
[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "black>=23.7.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["setuptools>=61.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["hillbot", "hillbot.core", "hillbot.services", "hillbot.utils", "hillbot.tests"]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
testpaths = ["hillbot/tests"]
python_files = "test_*.py"

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.isort]
profile = "black"
line_length = 100
